requirements.txt
SimpleITK
numpy==1.26.4
torch>=2.0.0
torchvision
timm
Pillow
numpy
albumentations
peft
safetensors
spams-bin